Massive rework to remove DPJSONSerializable; related style edits
This interface was extremely stateful, which made it tough to manage in classes
that had multiple outlets to having it activated.

We now split up the request/response cycle into explicit methods on each class
so that we can be certain of what context we shouldbe interpreting the response
in.